Making theme more dynamic.
In the tailwind utils I changed the theme config and module export to dynamically set the css classes based on the env BLOG_THEME or default. I also imported the THEME variable into _app.js to set the theme based on what was set in the tailwind util file. Next step might be to set up react context to allow a user to set their own theme.

Also, added more dummy posts to assist in making the pagination component.