Adding filter options that help a user narrow down a desired post.
Accordion component built and used to display filters for labels and authors. There is potential to consolidate the handling logic for the filters and only use one function, but I will look into that when I add the date as a filter option.

Displays an accordion that a user can click which reveals checkboxes. The user can select as many filters as they want, which dynamically updates the posts as they click.